% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_img2ascii_app}
\alias{launch_img2ascii_app}
\title{Launch the img2ASCII Shiny Application}
\usage{
launch_img2ascii_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp} (e.g., \code{launch.browser = TRUE}).}
}
\value{
Does not return a value; launches the Shiny application.
}
\description{
This function launches the Shiny application contained within the img2ASCII package.
}
